test(notebook-cloud): harden access smoke websocket harness
